% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCode.R
\docType{data}
\name{BWHCitationReport}
\alias{BWHCitationReport}
\title{WoS exported citation report for search AD=(brigham same anes*) OR AD=(brigham same anaes*)}
\description{
WoS exported citation report for search AD=(brigham same anes*) OR AD=(brigham same anaes*)
}
\keyword{data}

